move buildBlock2DateMap to utils
